コメントコントローラーをモジュール化
