+++ app/controllers/reports/comments_controller.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
-class Reports::CommentsController < CommentsController
+require_relative '../comments/comments_module'
+
+class Reports::CommentsController < ApplicationController
+  include Comments
+
   before_action :set_commentable
 
   private
